refactor: 기존 Widgets 컴포넌트를 shared로 이동, reportModal을 widgets로 이동
