+++ src/app/mocks/handlers/feed.ts
@@ -1,7 +1,7 @@
 import { http } from 'msw';
 
+import { feedMockData } from '../data';
 import { createHttpErrorResponse, createHttpSuccessResponse } from '../lib';
-
 interface IReportFeedReqDto {
   category: string;
   content: string;
@@ -16,6 +16,12 @@ export const feedHandlers = [
     if (!feedId || !category || !content) {
       return createHttpErrorResponse('피드 ID, 신고 카테고리, 컨텐츠가 다 입력되어야 합니다.');
     }
+
+    feedMockData.forEach((cur) => {
+      if (cur.id === +feedId) {
+        cur.isBlinded = true;
+      }
+    });
 
     return createHttpSuccessResponse({
       isReported: true,
