+++ src/entities/user/model/type.ts
@@ -1,0 +1,5 @@
+export interface IUser {
+  id: number;
+  profileImage: string;
+  username: string;
+}
