--- src/features/feed/reportModal/index.ts
@@ -1,2 +1,0 @@
-export * from './ui';
-export * from './lib';
