--- src/features/feed/reportModal/lib/index.ts
@@ -1,1 +1,0 @@
-export * from './mocks';
