--- src/features/feed/reportModal/lib/mocks/index.ts
@@ -1,1 +1,0 @@
-export * from './feed';
