+++ src/shared/types/image.ts
@@ -1,0 +1,4 @@
+export interface IImage {
+  id: number;
+  imageUrl: string;
+}
