+++ src/shared/ui/index.ts
@@ -4,3 +4,5 @@ export * from './button';
 export * from './modal-wrapper';
 export * from './checkbox';
 export * from './radio';
+export * from './modal';
+export * from './page-container';
