+++ src/shared/ui/modal/consts/index.ts
