+++ src/shared/ui/modal/hooks/index.ts
